Make throttling part of the core delivery service implementation
