+++ src/Demo/DemoDistributor/Endpoints/FileSystem/FileSystemDeliveryService.cs
@@ -1,17 +1,17 @@
 ﻿using System;
 using System.Threading.Tasks;
-using Delivered.Concurrency;
+using Delivered;
 
 namespace DemoDistributor.Endpoints.FileSystem
 {
-    public class FileSystemDeliveryService : ThrottledEndpointDeliveryService<File, FileSystemEndpoint>
+    public class FileSystemDeliveryService : EndpointDeliveryService<File, FileSystemEndpoint>
     {
         public FileSystemDeliveryService()
         {
             MaximumConcurrentDeliveries(5);
         }
 
-        protected override async Task DeliverThrottledAsync(File file, FileSystemEndpoint endpoint)
+        protected override async Task DeliverAsync(File file, FileSystemEndpoint endpoint)
         {
             Console.WriteLine($"Distributing file {file.Name} to File System directory {endpoint.Directory}");
